Refactor out resolver logic
